requirements.txt
# Web framework
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
websockets==14.1

# Config
python-dotenv==1.0.1
pyyaml==6.0.2

# Image processing
Pillow==11.1.0
numpy==1.26.4

# AI models
tensorflow==2.16.2
pytesseract==0.3.13
transformers==4.47.1
torch==2.5.1
torchvision==0.20.1
sentencepiece==0.2.0

# TTS
gTTS==2.5.4

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
httpx==0.28.1

# Linting
ruff==0.8.6
